+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Release to Github Packages
+- Explicitly flush datadog metrics after publishing them
+
+### Changed
+- Replace `statsd.batch` with direct calls, as it aggregates metrics interally by default now.
+  Also `#batch` method is deprecated and will be removed in version 6 of Datadog Statsd client.
 
 ## [0.3.1] - 2021-03-26
 ### Changed